Hm, warning on all implicit sign conversions was a bit much, changed to just warning on comparisons. VC++ only warns when it would affect the meaning of a comparison, however, so this still isn't right. I've added casts in any case to some equality comparisons (which now have subtly different and arguably wrong meanings, but it should make it easier to keep the warnings under control. Lexical tokens >2GB will just have to fail to work in a different manner)
--HG--
branch : Small bugs

+++ TameParse/Lr/lalr_builder.cpp
@@ -135,7 +135,7 @@ void lalr_builder::complete_parser() {
             int         offset  = (*item)->offset();
             
             // Items at the end of a rule don't produce any transitions
-            if (offset == rule.items().size()) continue;
+            if (offset == (int) rule.items().size()) continue;
             
             // Get the item that the 'dot' is before
             const item_container& dottedItem = rule.items()[offset];
